Add separate Draw component
